Scraper/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
attrs==21.4.0
    # via
    #   automat
    #   service-identity
    #   twisted
automat==22.10.0
    # via twisted
certifi==2021.10.8
    # via requests
cffi==1.15.1
    # via cryptography
chardet==4.0.0
    # via requests
constantly==15.1.0
    # via twisted
cryptography==3.3.2
    # via
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.1.0
    # via
    #   parsel
    #   scrapy
enum34==1.1.10
    # via cryptography

hyperlink==21.0.0
    # via twisted
idna==2.10
    # via
    #   hyperlink
    #   requests
    #   tldextract
incremental==22.10.0
    # via twisted
ipaddress==1.0.23
    # via
    #   cryptography
    #   service-identity
lxml==4.9.2
    # via
    #   parsel
    #   scrapy
parsel==1.6.0
    # via scrapy
pillow==6.2.2
    # via -r requirements.in
protego==0.2.1
    # via scrapy
pyasn1-modules==0.3.0
    # via service-identity
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   service-identity
pycparser==2.21
    # via cffi
pydispatcher==2.0.7
    # via scrapy
pyhamcrest==1.10.1
    # via twisted
pyopenssl==21.0.0
    # via scrapy
queuelib==1.6.1
    # via scrapy
requests-file==1.5.1
    # via tldextract
requests==2.27.1
    # via
    #   requests-file
    #   tldextract
scrapy==1.8.3
    # via -r requirements.in
service-identity==21.1.0
    # via scrapy
six==1.16.0
    # via
    #   automat
    #   cryptography
    #   parsel
    #   protego
    #   pyhamcrest
    #   pyopenssl
    #   requests-file
    #   scrapy
    #   service-identity
    #   w3lib
tldextract==2.2.3
    # via scrapy
twisted==20.3.0
    # via scrapy
typing==3.7.4.3
    # via hyperlink
urllib3==1.26.15
    # via requests
w3lib==1.22.0
    # via
    #   parsel
    #   scrapy
zope.interface==5.5.2
    # via
    #   scrapy
    #   twisted
azure-storage-blob==12.9.0
# The following packages are considered to be unsafe in a requirements file:
# setuptools
